+++ src/board.js
@@ -3,8 +3,8 @@
 	the squares it contains.
 */
 
-import { EventEmitter } from 'events';
 import { Piece, PieceType, SideType } from './piece';
+import { EventEmitter } from 'events';
 import { Square } from './square';
 
 // types
@@ -38,8 +38,8 @@ export class Board extends EventEmitter {
 	static create () {
 		let
 			b = new Board([]),
-			i = 0,
 			f = 0,
+			i = 0,
 			r = 0,
 			sq = null;
 
@@ -152,8 +152,8 @@ export class Board extends EventEmitter {
 
 	getSquares (side) {
 		let
-			list = [],
-			i = 0;
+			i = 0,
+			list = [];
 
 		for (i = 0; i < this.squares.length; i++) {
 			if (this.squares[i].piece &&