+++ test/common.js
@@ -1,4 +1,2 @@
 global.chai = require('chai');
 global.assert = chai.assert;
-global.expect = chai.expect;
-global.should = chai.should();