+++ test/src/boardValidation.js
@@ -3,10 +3,8 @@
 import { BoardValidation } from '../../src/boardValidation';
 import { Game } from '../../src/game';
 
-describe('BoardValidation', function() {
-	'use strict';
-
-	let getValidSquares = function(sq, validMoves) {
+describe('BoardValidation', () => {
+	let getValidSquares = (sq, validMoves) => {
 		let i = 0;
 
 		for (; i < validMoves.length; i++) {
@@ -17,34 +15,40 @@ describe('BoardValidation', function() {
 	};
 
 	// validate error creating BoardValidation when board is null
-	it('should fail if validation object is created without a valid board', function() {
+	it('should fail if validation object is created without a valid board', () => {
 		let bv = BoardValidation.create(null);
 
-		bv.start(function(err) {
+		bv.start((err) => {
 			assert.ok(err);
 			assert.strictEqual(err.message, 'board is invalid');
 		});
 	});
 
 	// ensure board and game set properly when BoardValidation is created
-	it('should properly reflect board and game when validation object is created', function() {
+	it('should properly reflect board and game when validation object is created', () => {
 		let
-			g = Game.create(),
-			b = g.board,
-			bv = BoardValidation.create(g);
+			b,
+			bv,
+			g = Game.create();
+
+		b = g.board;
+		bv = BoardValidation.create(g);
 
 		assert.strictEqual(bv.board, b);
 	});
 
 	// ensure validation returns appropriate piece move options based on turn
-	it('should properly indicate that a white Pawn has 2 valid moves', function() {
+	it('should properly indicate that a white Pawn has 2 valid moves', () => {
 		let
+			b,
+			bv,
 			g = Game.create(),
-			b = g.board,
-			bv = BoardValidation.create(g),
 			squares = [];
 
-		bv.start(function(err, validMoves) {
+		b = g.board;
+		bv = BoardValidation.create(g);
+
+		bv.start((err, validMoves) => {
 			if (err) {
 				throw err;
 			}
@@ -58,14 +62,17 @@ describe('BoardValidation', function() {
 	});
 
 	// ensure validation returns appropriate piece move options based on turn
-	it('testBoardValidation_BlackPawn_NoValidMoves', function() {
+	it('testBoardValidation_BlackPawn_NoValidMoves', () => {
 		let
+			b,
+			bv,
 			g = Game.create(),
-			b = g.board,
-			bv = BoardValidation.create(g),
 			squares = [];
 
-		bv.start(function(err, validMoves) {
+		b = g.board;
+		bv = BoardValidation.create(g);
+
+		bv.start((err, validMoves) => {
 			if (err) {
 				throw err;
 			}
@@ -79,22 +86,29 @@ describe('BoardValidation', function() {
 	});
 
 	// validate is square attacked on piece not being attacked
-	it('testBoardValidation_WhiteKing_IsNotAttacked', function() {
+	it('testBoardValidation_WhiteKing_IsNotAttacked', () => {
 		let
+			b,
+			bv,
 			g = Game.create(),
-			b = g.board,
-			bv = BoardValidation.create(g),
-			kingSquare = b.getSquare('e', 1);
+			kingSquare;
+
+		b = g.board;
+		bv = BoardValidation.create(g);
+		kingSquare = b.getSquare('e', 1);
 
 		assert.strictEqual(bv.isSquareAttacked(kingSquare), false);
 	});
 
 	// validate is square attacked on piece being attacked
-	it('testBoardValidation_BlackPawn_IsAttacked', function() {
+	it('testBoardValidation_BlackPawn_IsAttacked', () => {
 		let
-			g = Game.create(),
-			b = g.board,
-			bv = BoardValidation.create(g);
+			b,
+			bv,
+			g = Game.create();
+
+		b = g.board;
+		bv = BoardValidation.create(g);
 
 		b.move(b.getSquare('d', 2), b.getSquare('d', 4));
 		b.move(b.getSquare('d', 7), b.getSquare('d', 5));
@@ -104,11 +118,14 @@ describe('BoardValidation', function() {
 	});
 
 	// ensure is square attacked accurately tracks king being attacked
-	it('testBoardValidation_BlackKing_IsAttacked', function() {
+	it('testBoardValidation_BlackKing_IsAttacked', () => {
 		let
-			g = Game.create(),
-			b = g.board,
-			bv = BoardValidation.create(g);
+			b,
+			bv,
+			g = Game.create();
+
+		b = g.board;
+		bv = BoardValidation.create(g);
 
 		b.move(b.getSquare('e', 2), b.getSquare('e', 4));
 		b.move(b.getSquare('f', 7), b.getSquare('f', 5));
@@ -120,11 +137,14 @@ describe('BoardValidation', function() {
 	// ensure is square attacked accurately tracks that's it not being attacked
 	// based on bug found where board logic thought square was attacked by pawn more
 	// than 1 square away diagonally
-	it('testBoardValidation_BlackKing_IsNotAttacked', function() {
+	it('testBoardValidation_BlackKing_IsNotAttacked', () => {
 		let
-			g = Game.create(),
-			b = g.board,
-			bv = BoardValidation.create(g);
+			b,
+			bv,
+			g = Game.create();
+
+		b = g.board;
+		bv = BoardValidation.create(g);
 
 		b.move('g2', 'g4');
 		b.move('f7', 'f6');
@@ -136,18 +156,21 @@ describe('BoardValidation', function() {
 	});
 
 	// validate castle rule to left
-	it('testBoardValidation_WhiteKingCastle_Left', function() {
+	it('testBoardValidation_WhiteKingCastle_Left', () => {
 		let
+			b,
+			bv,
 			g = Game.create(),
-			b = g.board,
-			bv = BoardValidation.create(g),
 			squares = [];
+
+		b = g.board;
+		bv = BoardValidation.create(g);
 
 		b.getSquare('b', 1).piece = null;
 		b.getSquare('c', 1).piece = null;
 		b.getSquare('d', 1).piece = null;
 
-		bv.start(function(err, validMoves) {
+		bv.start((err, validMoves) => {
 			if (err) {
 				throw err;
 			}
@@ -162,19 +185,22 @@ describe('BoardValidation', function() {
 	});
 
 	// validate castle rule to right
-	it('testBoardValidation_BlackKingCastle_Right', function() {
+	it('testBoardValidation_BlackKingCastle_Right', () => {
 		let
+			b,
+			bv,
 			g = Game.create(),
-			b = g.board,
-			bv = BoardValidation.create(g),
 			squares = [];
+
+		b = g.board;
+		bv = BoardValidation.create(g);
 
 		b.getSquare('f', 8).piece = null;
 		b.getSquare('g', 8).piece = null;
 
 		b.move(b.getSquare('a', 2), b.getSquare('a', 4));
 
-		bv.start(function(err, validMoves) {
+		bv.start((err, validMoves) => {
 			if (err) {
 				throw err;
 			}
@@ -189,12 +215,15 @@ describe('BoardValidation', function() {
 	});
 
 	// validate castle rule no longer applies when king has moved
-	it('testBoardValidation_WhiteKingCastle_KingMoved', function() {
+	it('testBoardValidation_WhiteKingCastle_KingMoved', () => {
 		let
+			b,
+			bv,
 			g = Game.create(),
-			b = g.board,
-			bv = BoardValidation.create(g),
 			squares = [];
+
+		b = g.board;
+		bv = BoardValidation.create(g);
 
 		// clear squares between king and rook
 		b.getSquare('f', 1).piece = null;
@@ -204,7 +233,7 @@ describe('BoardValidation', function() {
 		b.move(b.getSquare('e', 1), b.getSquare('f', 1));
 		b.move(b.getSquare('f', 1), b.getSquare('e', 1));
 
-		bv.start(function(err, validMoves) {
+		bv.start((err, validMoves) => {
 			if (err) {
 				throw err;
 			}
@@ -218,12 +247,15 @@ describe('BoardValidation', function() {
 	});
 
 	// validate castle rule no longer applies when rook has moved
-	it('testBoardValidation_WhiteKingCastle_RookMoved', function() {
+	it('testBoardValidation_WhiteKingCastle_RookMoved', () => {
 		let
+			b,
+			bv,
 			g = Game.create(),
-			b = g.board,
-			bv = BoardValidation.create(g),
 			squares = [];
+
+		b = g.board;
+		bv = BoardValidation.create(g);
 
 		// clear squares between king and rook
 		b.getSquare('f', 1).piece = null;
@@ -233,7 +265,7 @@ describe('BoardValidation', function() {
 		b.move(b.getSquare('h', 1), b.getSquare('f', 1));
 		b.move(b.getSquare('f', 1), b.getSquare('h', 1));
 
-		bv.start(function(err, validMoves) {
+		bv.start((err, validMoves) => {
 			if (err) {
 				throw err;
 			}
@@ -247,22 +279,28 @@ describe('BoardValidation', function() {
 	});
 
 	// validate only move options are to block check
-	it('testBoardValidation_WhiteKingInCheck_BlockMovesOnly', function() {
+	it('testBoardValidation_WhiteKingInCheck_BlockMovesOnly', () => {
 		let
+			b,
+			bishopSquare,
+			bv,
 			g = Game.create(),
-			b = g.board,
-			bv = BoardValidation.create(g),
-			queenSquare = b.getSquare('d', 1),
-			bishopSquare = b.getSquare('f', 1),
-			knightSquare = b.getSquare('g', 1),
+			knightSquare,
+			queenSquare,
 			squares = [];
+
+		b = g.board;
+		bishopSquare = b.getSquare('f', 1);
+		bv = BoardValidation.create(g);
+		knightSquare = b.getSquare('g', 1);
+		queenSquare = b.getSquare('d', 1);
 
 		// put king in check
 		b.getSquare('e', 2).piece = null;
 		b.getSquare('e', 7).piece = null;
 		b.move(b.getSquare('d', 8), b.getSquare('e', 7), true);
 
-		bv.start(function(err, validMoves) {
+		bv.start((err, validMoves) => {
 			if (err) {
 				throw err;
 			}
@@ -280,12 +318,15 @@ describe('BoardValidation', function() {
 		});
 	});
 
-	it('should properly trigger game to emit check when King is placed in check', function () {
+	it('should properly trigger game to emit check when King is placed in check', () => {
 		let
-			g = Game.create(),
-			b = g.board,
-			bv = BoardValidation.create(g),
-			checkResult = [];
+			b,
+			bv,
+			checkResult = [],
+			g = Game.create();
+
+		b = g.board;
+		bv = BoardValidation.create(g);
 
 		// capture check event
 		g.on('check', (result) => (checkResult.push(result)));
@@ -308,7 +349,7 @@ describe('BoardValidation', function() {
 		// double-check (a checkmate... Queen and Knight both attacking)
 		b.move(b.getSquare('e', 4), b.getSquare('f', 6));
 
-		bv.start(function(err) {
+		bv.start((err) => {
 			if (err) {
 				throw err;
 			}
@@ -320,12 +361,15 @@ describe('BoardValidation', function() {
 	});
 
 	// validate inability to castle while king is in check
-	it('testBoardValidation_WhiteKingCastle_KingInCheck', function() {
+	it('testBoardValidation_WhiteKingCastle_KingInCheck', () => {
 		let
+			b,
+			bv,
 			g = Game.create(),
-			b = g.board,
-			bv = BoardValidation.create(g),
 			squares = [];
+
+		b = g.board;
+		bv = BoardValidation.create(g);
 
 		// clear squares between king and rook
 		b.getSquare('f', 1).piece = null;
@@ -336,7 +380,7 @@ describe('BoardValidation', function() {
 		b.getSquare('e', 7).piece = null;
 		b.move(b.getSquare('d', 8), b.getSquare('e', 7), true);
 
-		bv.start(function(err, validMoves) {
+		bv.start((err, validMoves) => {
 			if (err) {
 				throw err;
 			}
@@ -350,12 +394,15 @@ describe('BoardValidation', function() {
 	});
 
 	// validate inability to castle through or into check
-	it('testBoardValidation_WhiteKingCastle_MoveThroughCheck', function() {
+	it('testBoardValidation_WhiteKingCastle_MoveThroughCheck', () => {
 		let
+			b,
+			bv,
 			g = Game.create(),
-			b = g.board,
-			bv = BoardValidation.create(g),
 			squares = [];
+
+		b = g.board;
+		bv = BoardValidation.create(g);
 
 		// clear squares between king and rook
 		b.getSquare('f', 1).piece = null;
@@ -366,7 +413,7 @@ describe('BoardValidation', function() {
 		b.getSquare('e', 7).piece = null;
 		b.move(b.getSquare('d', 8), b.getSquare('f', 6), true);
 
-		bv.start(function(err, validMoves) {
+		bv.start((err, validMoves) => {
 			if (err) {
 				throw err;
 			}
@@ -380,12 +427,15 @@ describe('BoardValidation', function() {
 	});
 
 	// validate inability to move into check
-	it('testBoardValidation_WhiteKing_UnableToExposeCheck', function() {
+	it('testBoardValidation_WhiteKing_UnableToExposeCheck', () => {
 		let
+			b,
+			bv,
 			g = Game.create(),
-			b = g.board,
-			bv = BoardValidation.create(g),
 			squares = [];
+
+		b = g.board;
+		bv = BoardValidation.create(g);
 
 		// block king with knight and attack with queen
 		b.getSquare('e', 2).piece = null;
@@ -393,7 +443,7 @@ describe('BoardValidation', function() {
 		b.move(b.getSquare('d', 8), b.getSquare('e', 7), true);
 		b.move(b.getSquare('g', 1), b.getSquare('e', 2), true);
 
-		bv.start(function(err, validMoves) {
+		bv.start((err, validMoves) => {
 			if (err) {
 				throw err;
 			}
@@ -407,11 +457,14 @@ describe('BoardValidation', function() {
 	});
 
 	// validate checkmate (no available moves)
-	it('testBoardValidation_BlackKing_Checkmate', function() {
+	it('testBoardValidation_BlackKing_Checkmate', () => {
 		let
-			g = Game.create(),
-			b = g.board,
-			bv = BoardValidation.create(g);
+			b,
+			bv,
+			g = Game.create();
+
+		b = g.board;
+		bv = BoardValidation.create(g);
 
 		// put king into checkmate
 		b.move(b.getSquare('e', 2), b.getSquare('e', 4));
@@ -420,7 +473,7 @@ describe('BoardValidation', function() {
 		b.move(b.getSquare('g', 7), b.getSquare('g', 5));
 		b.move(b.getSquare('d', 1), b.getSquare('h', 5));
 
-		bv.start(function(err, validMoves) {
+		bv.start((err, validMoves) => {
 			if (err) {
 				throw err;
 			}
@@ -430,24 +483,27 @@ describe('BoardValidation', function() {
 	});
 
 	// validate pieces don't disappear after validation
-	it('testBoardValidation_Pawn_Disappears', function() {
+	it('testBoardValidation_Pawn_Disappears', () => {
 		let
-			g = Game.create(),
-			b = g.board,
-			bv = BoardValidation.create(g);
+			b,
+			bv,
+			g = Game.create();
+
+		b = g.board;
+		bv = BoardValidation.create(g);
 
 		b.move('e2', 'e4');
 		b.move('e7', 'e6');
 		b.move('d2', 'd4');
 		b.move('d7', 'd5');
 
-		bv.start(function() {
+		bv.start(() => {
 			assert.ok(b.getSquare('d4').piece !== null);
 		});
 
 		b.move('b1', 'c3');
 
-		bv.start(function() {
+		bv.start(() => {
 			assert.ok(b.getSquare('d4').piece !== null, 'pawn has disappeared during validation');
 		});
 	});