+++ test/src/gameValidation.js
@@ -3,15 +3,15 @@
 import { Game } from '../../src/game';
 import { GameValidation } from '../../src/gameValidation';
 
-describe('GameValidation', function() {
-	'use strict';
-
+describe('GameValidation', () => {
 	// validate check
-	it('should properly indicate check', function() {
+	it('should properly indicate check', () => {
 		let
+			b,
 			g = Game.create(),
-			b = g.board,
 			v = GameValidation.create(g);
+
+		b = g.board;
 
 		// put king into check
 		b.move(b.getSquare('d', 2), b.getSquare('d', 4));
@@ -19,7 +19,7 @@ describe('GameValidation', function() {
 		b.move(b.getSquare('d', 4), b.getSquare('e', 5));
 		b.move(b.getSquare('f', 8), b.getSquare('b', 4));
 
-		v.start(function(err, result) {
+		v.start((err, result) => {
 			if (err) {
 				throw err;
 			}
@@ -32,15 +32,17 @@ describe('GameValidation', function() {
 	});
 
 	// validate Knight check (part 1)
-	it('should properly indicate check due to Knight (part 1)', function () {
+	it('should properly indicate check due to Knight (part 1)', () => {
 		let
+			b,
 			g = Game.create(),
-			b = g.board,
 			v = GameValidation.create(g);
+
+		b = g.board;
 
 		b.move(b.getSquare('b1'), b.getSquare('c7'));
 
-		v.start(function (err, result) {
+		v.start((err, result) => {
 			if (err) {
 				throw err;
 			}
@@ -50,11 +52,13 @@ describe('GameValidation', function() {
 	});
 
 	// validate Knight check (part 2)
-	it('should properly indicate check due to Knight (part 2)', function () {
+	it('should properly indicate check due to Knight (part 2)', () => {
 		let
+			b,
 			g = Game.create(),
-			b = g.board,
 			v = GameValidation.create(g);
+
+		b = g.board;
 
 		b.move(b.getSquare('b1'), b.getSquare('c3'));
 		b.move(b.getSquare('g8'), b.getSquare('f6'));
@@ -62,7 +66,7 @@ describe('GameValidation', function() {
 		b.move(b.getSquare('f6'), b.getSquare('g8'));
 		b.move(b.getSquare('b5'), b.getSquare('d6'));
 
-		v.start(function (err, result) {
+		v.start((err, result) => {
 			if (err) {
 				throw err;
 			}
@@ -72,11 +76,13 @@ describe('GameValidation', function() {
 	});
 
 	// validate checkmate
-	it('should properly indicate checkmate', function() {
+	it('should properly indicate checkmate', () => {
 		let
+			b,
 			g = Game.create(),
-			b = g.board,
 			v = GameValidation.create(g);
+
+		b = g.board;
 
 		// put king into checkmate
 		b.move(b.getSquare('e', 2), b.getSquare('e', 4));
@@ -85,7 +91,7 @@ describe('GameValidation', function() {
 		b.move(b.getSquare('g', 7), b.getSquare('g', 5));
 		b.move(b.getSquare('d', 1), b.getSquare('h', 5));
 
-		v.start(function(err, result) {
+		v.start((err, result) => {
 			if (err) {
 				throw err;
 			}
@@ -99,11 +105,13 @@ describe('GameValidation', function() {
 	});
 
 	// validate stalemate
-	it('should properly indicate stalemate', function() {
+	it('should properly indicate stalemate', () => {
 		let
+			b,
 			g = Game.create(),
-			b = g.board,
 			v = GameValidation.create(g);
+
+		b = g.board;
 
 		// remove several pieces from the board
 		b.getSquare('a1').piece = null;
@@ -140,7 +148,7 @@ describe('GameValidation', function() {
 		b.move('e1', 'e6', true);
 		b.move('e2', 'e7');
 
-		v.start(function(err, result) {
+		v.start((err, result) => {
 			if (err) {
 				throw err;
 			}
@@ -155,11 +163,13 @@ describe('GameValidation', function() {
 
 	// validate threefold repetition rule
 	// Fischer vs Petrosian, Buenos Aires, 1971, round 3
-	it('should properly indicate 3-fold repetition', function() {
+	it('should properly indicate 3-fold repetition', () => {
 		let
+			b,
 			g = Game.create(),
-			b = g.board,
 			v = GameValidation.create(g);
+
+		b = g.board;
 
 		// 1. e4 e6
 		b.move('e2', 'e4');
@@ -263,7 +273,7 @@ describe('GameValidation', function() {
 		// 34. Qe2
 		b.move('d3', 'e2');
 
-		v.start(function(err, result) {
+		v.start((err, result) => {
 			if (err) {
 				throw err;
 			}